Updated JAXB version and split up Jaxb classes
