+++ pom.xml
@@ -42,6 +42,7 @@
         <apache.commons.lang3.version>3.4</apache.commons.lang3.version>
         <junit.version>4.12</junit.version>
         <gson.version>2.8.5</gson.version>
+        <jaxb.version>2.3.2</jaxb.version>
     </properties>
 
     <dependencyManagement>
@@ -180,7 +181,7 @@
                 <plugin>
                     <groupId>org.apache.felix</groupId>
                     <artifactId>maven-bundle-plugin</artifactId>
-                    <version>3.5.0</version>
+                    <version>4.2.0</version>
                     <extensions>true</extensions>
                     <configuration>
                         <!-- Disabled OBR to increase build speed -->
