+++ standards/genc/src/main/java/org/codice/countrycode/standards/genc/Division.java
@@ -1,0 +1,45 @@
+/**
+ * Copyright (c) Codice Foundation
+ *
+ * <p>This is free software: you can redistribute it and/or modify it under the terms of the GNU
+ * Lesser General Public License as published by the Free Software Foundation, either version 3 of
+ * the License, or any later version.
+ *
+ * <p>This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public
+ * License is distributed along with this program and can be found at
+ * <http://www.gnu.org/licenses/lgpl.html>.
+ */
+package org.codice.countrycode.standards.genc;
+
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlAttribute;
+import javax.xml.bind.annotation.XmlRootElement;
+import javax.xml.bind.annotation.XmlValue;
+
+@XmlRootElement
+@XmlAccessorType(XmlAccessType.FIELD)
+public class Division {
+
+  @XmlAttribute private String codeSpace;
+
+  @XmlValue private String content;
+
+  public void setCodeSpace(String codeSpace) {
+    this.codeSpace = codeSpace;
+  }
+
+  public void setContent(String content) {
+    this.content = content;
+  }
+
+  public String getCodeSpace() {
+    return codeSpace;
+  }
+
+  public String getContent() {
+    return content;
+  }
+}
