+++ standards/genc/src/main/java/org/codice/countrycode/standards/genc/LocalShortName.java
@@ -1,0 +1,44 @@
+/**
+ * Copyright (c) Codice Foundation
+ *
+ * <p>This is free software: you can redistribute it and/or modify it under the terms of the GNU
+ * Lesser General Public License as published by the Free Software Foundation, either version 3 of
+ * the License, or any later version.
+ *
+ * <p>This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public
+ * License is distributed along with this program and can be found at
+ * <http://www.gnu.org/licenses/lgpl.html>.
+ */
+package org.codice.countrycode.standards.genc;
+
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlRootElement;
+
+@XmlRootElement
+@XmlAccessorType(XmlAccessType.FIELD)
+public class LocalShortName {
+
+  @XmlElement private String name;
+
+  public void setName(String name) {
+    this.name = name;
+  }
+
+  public void setIso6393Char3Code(String iso6393Char3Code) {
+    this.iso6393Char3Code = iso6393Char3Code;
+  }
+
+  private String iso6393Char3Code;
+
+  public String getIso6393Char3Code() {
+    return iso6393Char3Code;
+  }
+
+  public String getName() {
+    return name;
+  }
+}
